giovi-ai/agency-service/pyproject.toml
[project]
name = "agency-service"
version = "0.1.0"
description = "Modulo FastAPI per la gestione delle agenzie di pulizie di Giovi AI"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Giovi AI" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-cloud-firestore>=2.16.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "anyio>=4.4.0",
  "ruff>=0.6.9"
]

[project.urls]
repository = "https://github.com/giovi-ai/giovi-ai/tree/main/giovi-ai/agency-service"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = [
  "src"
]

